refactor: 가격 필드 Int 형으로 변경 - 부동 소수점은 연산시에 오차가 생기기 때문에 정수형으로 사용 - 13.33$ 와 같이 소수점이 필요한 경우 단위를 센트로 사용(1333)

+++ src/main/kotlin/com/chaw/shopping_note/app/receipt/application/dto/CreateReceiptItemRequestDto.kt
@@ -7,7 +7,7 @@ data class CreateReceiptItemRequestDto(
     val receiptId: Long,
     val productName: String,
     val productCode: String,
-    val unitPrice: Double,
+    val unitPrice: Int,
     val quantity: Int,
     val category: Category
 )
